requirements.txt
msgpack==1.1.0
packaging==24.2
pillow==11.1.0
pyserial==3.5
python-can==4.5.0
ttkthemes==3.2.2
typing_extensions==4.12.2
wrapt==1.17.2
pygame
